+++ src/main/java/com/meetkey/server/domain/chat/controller/ChatController.java
@@ -1,14 +1,8 @@
 package com.meetkey.server.domain.chat.controller;
 
-import com.meetkey.server.domain.chat.dto.pub.ChatMessagePubDTO;
-import com.meetkey.server.domain.chat.dto.request.ChatMessageSendReqDTO;
 import com.meetkey.server.domain.chat.dto.request.ChatReqDTO;
-import com.meetkey.server.domain.chat.dto.response.ChatMessageResDTO;
 import com.meetkey.server.domain.chat.dto.response.ChatResDTO;
-import com.meetkey.server.domain.chat.entity.ChatMessage;
-import com.meetkey.server.domain.chat.message.redis.ChatRedisPublisher;
 import com.meetkey.server.domain.chat.service.command.ChatCommandService;
-import com.meetkey.server.domain.chat.service.command.ChatMessageCommandService;
 import com.meetkey.server.domain.chat.service.query.ChatQueryService;
 import com.meetkey.server.global.apiPayload.response.BasicResponse;
 import com.meetkey.server.global.apiPayload.status.CommonSuccessStatus;
@@ -16,25 +10,19 @@ import com.meetkey.server.global.security.CustomUserDetails;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
-import org.springframework.messaging.handler.annotation.MessageMapping;
-import org.springframework.messaging.handler.annotation.Payload;
-import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
-import java.security.Principal;
 import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
-@Tag(name = "채팅  API")
+@Tag(name = "채팅 API")
 @RequestMapping("/chat-room")
 public class ChatController {
 
     private final ChatCommandService chatCommandService;
     private final ChatQueryService chatQueryService;
-    private final ChatMessageCommandService chatMessageCommandService;
-    private final ChatRedisPublisher chatRedisPublisher;
 
     @Operation(summary = "채팅방 생성 API by 제인", description = "채팅방을 생성하는 API by 제인")
     @PostMapping
@@ -84,23 +72,5 @@ public class ChatController {
     ) {
         chatCommandService.readMessages(details.getMemberId(), chatRoomId);
         return BasicResponse.success(CommonSuccessStatus._OK, null);
-    }
-
-    @MessageMapping("/chat/send")
-    public void sendMessage(@Payload ChatMessageSendReqDTO req, Principal principal) {
-        CustomUserDetails details = (CustomUserDetails) principal;
-        Long senderId = details.getMemberId();
-
-        ChatMessage message = chatMessageCommandService.sendMessage(
-                senderId,
-                req.getChatRoomId(),
-                req.getMessageType(),
-                req.getContent(),
-                req.getMediaUrl(),
-                req.getDuration() // VOICE 메시지일 때만 사용, TEXT와 IMAGE는 null
-        );
-
-        // 같은 채팅방 구독자에게 Redis로 publish
-        chatRedisPublisher.publish(ChatMessagePubDTO.from(message));
     }
 }
